Refactor AbstractProvider and GoogleProvider classes; add YandexProvider implementation and remove old Yandex class
